fix(sentinel): 修正 BlockException 处理方法签名与日志资源名
- 调整 EagleSentinelBlockExceptionHandler.handle 方法参数，新增 resourceName 字符串参数
- 修改日志中 BlockException 的资源名称输出为传入的 resourceName
- 保持异常处理逻辑和响应状态码不变

feat(seata): 重构 Seata 配置环境注入方式为 ApplicationListener

- 将 SeataEnvironmentPostProcessor 从 EnvironmentPostProcessor 改为 ApplicationListener<ApplicationEnvironmentPreparedEvent>
- 修改配置同步逻辑以适应新的事件监听接口
- 在 spring.factories 中注册新的 ApplicationListener 实现
- 保证配置属性注入优先级符合预期

test(redis): 添加 CacheProtectionUtil 单元测试覆盖

- 实现对 getWithMutex 方法的缓存击穿保护功能测试
- 覆盖缓存命中、缓存穿透、加锁流程和空值占位符处理场景
- 编写 jitter 方法的随机偏移和边界测试
- 使用 Mockito 模拟 Redis 和 Redisson 行为进行隔离测试

build(sentinel): 添加 jakarta.servlet-api 依赖为编译时依赖

- 在 eagle-sentinel-starter 的 build.gradle 中增加 compileOnly jakarta.servlet-api 依赖，支持编译环境使用但不打包

+++ eagle-starter/eagle-rocketmq-starter/src/main/java/com/eagle/rocketmq/properties/RocketMqProperties.java
@@ -58,7 +58,7 @@ public class RocketMqProperties {
          * 本地缓存消息总字节数上限（字节），默认 64 MB。
          * 与 {@link #maxCachedMessageCount} 同时生效，任意一个触发则暂停拉取。
          */
-        private long maxCachedMessageSizeInBytes = 64 * 1024 * 1024L;
+        private int maxCachedMessageSizeInBytes = 64 * 1024 * 1024;
 
         /**
          * 重试次数告警阈值。
